+++ app/api/github-search/route.ts
@@ -1,8 +1,9 @@
 import { NextRequest, NextResponse } from "next/server"
-import { fetchGrowthStats } from "@/lib/utils"
+import { subDays, formatISO } from 'date-fns'
 
 const GITHUB_GRAPHQL_URL = "https://api.github.com/graphql"
 const GITHUB_REST_SEARCH_URL = "https://api.github.com/search/code"
+const GITHUB_API = 'https://api.github.com'
 
 function getGithubToken() {
   return process.env.GITHUB_TOKEN || process.env.NEXT_PUBLIC_GITHUB_TOKEN
@@ -22,6 +23,7 @@ async function fetchGraphQLRepos(keywords: string[]): Promise<any[]> {
             owner { login }
             createdAt
             pushedAt
+            isArchived  
             repositoryTopics(first: 10) {
               nodes { topic { name } }
             }
@@ -113,31 +115,6 @@ async function fetchCodeMatches(keywords: string[]): Promise<any[]> {
   return allMatches
 }
 
-/* function mergeResults(repos: any[], codeMatches: any[]): any[] {
-  // Map code matches by repo full_name
-  const codeByRepo: Record<string, any[]> = {}
-  for (const match of codeMatches) {
-    const repoName = match.repository.full_name
-    if (!codeByRepo[repoName]) codeByRepo[repoName] = []
-    codeByRepo[repoName].push({
-      file_name: match.name,
-      file_url: match.html_url,
-      snippet: match.text_matches?.[0]?.fragment || "",
-    })
-  }
-  // Merge code matches into repos
-  return repos.map((repo) => {
-    const fullName = `${repo.owner.login}/${repo.name}`
-    return {
-      repo_name: repo.name,
-      repo_url: repo.url,
-      stars: repo.stargazerCount,
-      topics: repo.repositoryTopics.nodes.map((n: any) => n.topic.name),
-      readme_content: repo.object?.text || "",
-      code_matches: codeByRepo[fullName] || [],
-    }
-  })
-} */
 
 
 function mergeResults(repos: any[], codeMatches: any[]): any[] {
@@ -157,6 +134,7 @@ function mergeResults(repos: any[], codeMatches: any[]): any[] {
         readme_content: repo.object?.text || "",
         createdAt: repo.createdAt,
         pushed_at: repo.pushedAt,
+        archived: repo.isArchived,
         code_matches: [],
       }
     }
@@ -185,6 +163,7 @@ function mergeResults(repos: any[], codeMatches: any[]): any[] {
           readme_content: match.repository.readme_content || "",
           createdAt: match.repository.createdAt || null,
           pushed_at: match.repository.pushedAt || null,
+          archived:false,
           code_matches: [codeItem],
         }
       }
@@ -214,6 +193,43 @@ async function fetchContributorsCount(owner: string, repo: string): Promise<numb
   // If no Link header, just count the returned contributors
   const data = await res.json()
   return Array.isArray(data) ? data.length : 0
+}
+
+export async function fetchGrowthStats(owner: string, repo: string) {
+  const headers = {
+    Authorization: `Bearer ${getGithubToken()}`,
+    Accept: 'application/vnd.github+json',
+  }
+
+  const sinceDate = formatISO(subDays(new Date(), 180))
+
+  const [commitsRes, issuesRes, contributorsRes, starsRes] = await Promise.all([
+    fetch(`${GITHUB_API}/repos/${owner}/${repo}/commits?since=${sinceDate}`, { headers }),
+    fetch(`${GITHUB_API}/repos/${owner}/${repo}/issues?since=${sinceDate}&state=closed`, { headers }),
+    fetch(`${GITHUB_API}/repos/${owner}/${repo}/contributors`, { headers }),
+    fetch(`${GITHUB_API}/repos/${owner}/${repo}/stargazers`, { headers }),
+  ])
+
+  const commits = await commitsRes.json()
+  const issues = await issuesRes.json()
+  const contributors = await contributorsRes.json()
+  const stars = await starsRes.json()
+
+  const newContributors = contributors.filter((c: any) => new Date(c.created_at) > new Date(sinceDate))
+
+  const growthScore =
+    commits.length * 0.3 +
+    newContributors.length * 1.0 +
+    issues.length * 0.5 +
+    stars.length * 0.2
+
+  return {
+    commits_last_month: commits.length,
+    new_contributors_last_month: newContributors.length,
+    issues_closed_last_month: issues.length,
+    stars_gained_last_month: stars.length,
+    growth_score: Math.round(growthScore * 10) / 10,
+  }
 }
 
 export async function POST(req: NextRequest) {
