+++ app/layout.tsx
@@ -2,7 +2,7 @@ import type React from "react"
 import type { Metadata } from "next"
 import { Inter } from "next/font/google"
 import "./globals.css"
-import { ThemeProvider } from "@/components/theme-provider"
+import { ThemeProvider } from "@/components/theme/ThemeProvider"
 import { ProductProvider } from '@/components/context/ProductContext'
 
 const inter = Inter({ subsets: ["latin"] })
