+++ app/page.tsx
@@ -12,23 +12,19 @@ import {
   Search,
   Plus,
   X,
-  ExternalLink,
-  TrendingUp,
   Sparkles,
   Target,
   ArrowRight,
-  Star,
-  Users,
-  Calendar,
-  ArrowLeft,
 } from "lucide-react"
-import Link from "next/link"
 import { useRouter } from "next/navigation"
-import Image from "next/image"
 import { type Product, transformGitHubToProduct, type GitHubRepository } from "@/lib/data"
-import { Header } from "@/components/header"
+import { Header } from "@/components/layout/Header"
 import { useProductContext } from '@/components/context/ProductContext'
 import { ProductList } from '@/components/product/ProductList'
+import LogoCircle from '@/components/common/LogoCircle'
+import HeroSection from '@/components/home/HeroSection'
+import ProductFormCard from '@/components/home/ProductFormCard'
+import SimilarProductsSection from '@/components/home/SimilarProductsSection'
 
 export default function HomePage() {
   const {
@@ -42,6 +38,8 @@ export default function HomePage() {
     setLastProductName,
     lastDescription,
     setLastDescription,
+    currentPage,
+    setCurrentPage,
   } = useProductContext();
 
   const router = useRouter();
@@ -50,14 +48,12 @@ export default function HomePage() {
   const [currentFeature, setCurrentFeature] = useState("")
   const [isSearched, setIsSearched] = useState(false)
   const [isLoading, setIsLoading] = useState(false)
-  const [currentPage, setCurrentPage] = useState(1)
   const [itemsPerPage] = useState(12)
   const [growthStats, setGrowthStats] = useState<Record<string, any>>({});
 
   const totalPages = Math.ceil(products.length / itemsPerPage)
   const startIndex = (currentPage - 1) * itemsPerPage
   const endIndex = startIndex + itemsPerPage
-  const currentResults = products.slice(startIndex, endIndex)
 
   useEffect(() => {
     setCurrentCompareProduct(null)
@@ -134,44 +130,6 @@ export default function HomePage() {
     setLastDescription(e.target.value);
   };
 
-  /* const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault()
-    if (productName.trim() && userFeatures.length > 0) {
-      setIsLoading(true)
-      setLastProductName(productName)
-      setLastDescription(description)
-      try {
-        const response = await fetch("/api/github-search", {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify({
-            description,
-            features: userFeatures,
-          }),
-        })
-        const data = await response.json()
-        if (data.repositories && Array.isArray(data.repositories)) {
-          const transformedProducts = data.repositories
-            .map((repo: GitHubRepository) => transformGitHubToProduct(repo, userFeatures))
-            .sort((a: Product, b: Product) => b.similarity - a.similarity)
-          setProducts(transformedProducts)
-          setUserFeatures(userFeatures)
-          setIsSearched(true)
-          setCurrentPage(1)
-        } else {
-          setProducts([])
-          setIsSearched(false)
-        }
-      } catch (error) {
-        setProducts([])
-        setIsSearched(false)
-      } finally {
-        setIsLoading(false)
-      }
-    }
-  } */
 
     const handleSubmit = async (e: React.FormEvent) => {
       e.preventDefault();
@@ -237,180 +195,47 @@ export default function HomePage() {
     
 
 
-  const handleKeyPress = (e: React.KeyboardEvent) => {
+  const handleFeatureKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
     if (e.key === "Enter") {
-      e.preventDefault()
-      addFeature()
+      e.preventDefault();
+      addFeature();
     }
-  }
+  };
 
   return (
-    <div className="min-h-screen bg-gradient-to-br from-slate-50 via-white to-indigo-50 dark:from-slate-950 dark:via-slate-900 dark:to-indigo-950">
+    <div className="min-h-screen relative bg-gradient-to-br from-slate-50 via-white to-indigo-50 dark:from-slate-950 dark:via-slate-900 dark:to-indigo-950">
       <Header />
-
+    
       <div className="container mx-auto px-4 py-8">
-        <div className="max-w-6xl mx-auto">
-          {/* Hero Section */}
-          <div className="text-center mb-12">
-            <div className="inline-flex items-center gap-2 bg-indigo-100 dark:bg-indigo-900/30 text-indigo-700 dark:text-indigo-300 px-4 py-2 rounded-full text-sm font-medium mb-6">
-              <Sparkles className="h-4 w-4" />
-              Productive AI - Product Analysis
-            </div>
-            <h1 className="text-5xl md:text-6xl font-bold bg-gradient-to-r from-slate-900 via-indigo-900 to-slate-900 dark:from-slate-100 dark:via-indigo-100 dark:to-slate-100 bg-clip-text text-transparent mb-6">
-              Discover Similar Products
-            </h1>
-            <p className="text-xl text-slate-600 dark:text-slate-400 max-w-2xl mx-auto leading-relaxed">
-              Find competing products, analyze feature overlap, and get insights to build better products
-            </p>
-          </div>
-
-          {/* Main Form Card */}
-          <Card className="mb-12 border-0 shadow-2xl bg-white/80 dark:bg-slate-900/80 backdrop-blur-sm">
-            <CardHeader className="pb-6">
-              <div className="flex items-center gap-3 mb-2">
-                <div className="h-10 w-10 bg-gradient-to-br from-indigo-500 to-purple-600 rounded-xl flex items-center justify-center">
-                  <Target className="h-5 w-5 text-white" />
-                </div>
-                <div>
-                  <CardTitle className="text-2xl">Analyze Your Product</CardTitle>
-                  <CardDescription className="text-base">
-                    Enter your product details to discover similar products and competitive insights
-                  </CardDescription>
-                </div>
-              </div>
-            </CardHeader>
-            <CardContent className="space-y-8">
-              <form onSubmit={handleSubmit} className="space-y-8">
-                <div className="grid grid-cols-1 lg:grid-cols-2 gap-8">
-                  <div className="space-y-6">
-                    <div className="space-y-3">
-                      <Label htmlFor="productName" className="text-base font-medium">
-                        Product Name
-                      </Label>
-                      <Input
-                        id="productName"
-                        placeholder="e.g., TimeTrackPro"
-                        value={productName}
-                        onChange={handleProductNameChange}
-                        className="h-12 text-base"
-                        required
-                      />
-                    </div>
-
-                    <div className="space-y-3">
-                      <Label htmlFor="description" className="text-base font-medium">
-                        Description
-                      </Label>
-                      <Textarea
-                        id="description"
-                        placeholder="Brief description of your product and its main purpose..."
-                        value={description}
-                        onChange={handleDescriptionChange}
-                        rows={4}
-                        className="text-base resize-none"
-                      />
-                    </div>
-                  </div>
-
-                  <div className="space-y-6">
-                    <div className="space-y-3">
-                      <Label htmlFor="features" className="text-base font-medium">
-                        Add Features
-                      </Label>
-                      <div className="flex gap-2">
-                        <Input
-                          id="features"
-                          placeholder="e.g., calendar, chat, time tracking"
-                          value={currentFeature}
-                          onChange={(e) => setCurrentFeature(e.target.value)}
-                          onKeyPress={handleKeyPress}
-                          className="h-12 text-base"
-                        />
-                        <Button
-                          type="button"
-                          onClick={addFeature}
-                          size="icon"
-                          className="h-12 w-12 bg-gradient-to-r from-indigo-500 to-purple-600 hover:from-indigo-600 hover:to-purple-700"
-                        >
-                          <Plus className="h-5 w-5" />
-                        </Button>
-                      </div>
-                    </div>
-
-                    {userFeatures.length > 0 && (
-                      <div className="space-y-3">
-                        <Label className="text-base font-medium">Features ({userFeatures.length})</Label>
-                        <div className="flex flex-wrap gap-2 p-4 bg-slate-50 dark:bg-slate-800/50 rounded-lg border-2 border-dashed border-slate-200 dark:border-slate-700">
-                          {userFeatures.map((feature) => (
-                            <Badge
-                              key={feature}
-                              className="flex items-center gap-2 px-3 py-2 text-sm bg-indigo-100 text-indigo-800 hover:bg-indigo-200 dark:bg-indigo-900/50 dark:text-indigo-200 dark:hover:bg-indigo-900/70"
-                            >
-                              {feature}
-                              <button
-                                type="button"
-                                onClick={() => removeFeature(feature)}
-                                className="hover:bg-indigo-200 dark:hover:bg-indigo-800 rounded-full p-0.5 transition-colors"
-                              >
-                                <X className="h-3 w-3" />
-                              </button>
-                            </Badge>
-                          ))}
-                        </div>
-                      </div>
-                    )}
-                  </div>
-                </div>
-
-                <Button
-                  type="submit"
-                  className="w-full h-14 text-lg font-semibold bg-gradient-to-r from-indigo-500 to-purple-600 hover:from-indigo-600 hover:to-purple-700 shadow-lg hover:shadow-xl transition-all duration-200"
-                  disabled={!productName.trim() || userFeatures.length === 0 || isLoading}
-                >
-                  {isLoading ? (
-                    <>
-                      <div className="animate-spin rounded-full h-5 w-5 border-2 border-white border-t-transparent mr-3" />
-                      Analyzing...
-                    </>
-                  ) : (
-                    <>
-                      <Search className="h-5 w-5 mr-3" />
-                      Find Similar Products
-                      <ArrowRight className="h-5 w-5 ml-3" />
-                    </>
-                  )}
-                </Button>
-              </form>
-            </CardContent>
-          </Card>
-
-          {/* Results Section */}
-          {isSearched && products.length === 0 && (
-            <div className="text-center text-lg text-slate-500 dark:text-slate-400 py-12">
-              No similar products found. Try different features or keywords.
-            </div>
-          )}
-          {isSearched && products.length > 0 && (
-            <div className="space-y-8">
-              <div className="text-center">
-                <h2 className="text-3xl font-bold text-slate-900 dark:text-slate-100 mb-2">Similar Products Found</h2>
-                <p className="text-slate-600 dark:text-slate-400">
-                  {products.length} products similar to "{productName}" based on feature analysis
-                </p>
-              </div>
-              <ProductList
-                products={productsWithGrowth}
-                currentPage={currentPage}
-                itemsPerPage={itemsPerPage}
-                goToPage={goToPage}
-                nextPage={nextPage}
-                prevPage={prevPage}
-                totalPages={totalPages}
-                startIndex={startIndex}
-                endIndex={endIndex}
-              />
-            </div>
-          )}
+        <div className="max-w-6xl mx-auto relative">
+          <HeroSection />
+          <ProductFormCard
+            productName={productName}
+            setProductName={setProductName}
+            description={description}
+            setDescription={setDescription}
+            currentFeature={currentFeature}
+            setCurrentFeature={setCurrentFeature}
+            userFeatures={userFeatures}
+            addFeature={addFeature}
+            removeFeature={removeFeature}
+            handleSubmit={handleSubmit}
+            isLoading={isLoading}
+            handleFeatureKeyDown={handleFeatureKeyDown}
+          />
+          <SimilarProductsSection
+            isSearched={isSearched}
+            products={productsWithGrowth}
+            productName={productName}
+            currentPage={currentPage}
+            itemsPerPage={itemsPerPage}
+            goToPage={goToPage}
+            nextPage={nextPage}
+            prevPage={prevPage}
+            totalPages={totalPages}
+            startIndex={startIndex}
+            endIndex={endIndex}
+          />
         </div>
       </div>
     </div>
